backend/geomina/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "alejandro", email = "alejandro.salas@live.cl"},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.1",
    "mysql>=0.0.3",
    "mysqlclient>=2.2.4",
    "django-cors-headers>=4.3.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
